Extract AddFundVendorModel to a separate file
