+++ lib/bloc/account/add_fund_vendor_model.dart
@@ -1,0 +1,11 @@
+class AddFundVendorModel {
+  String name;
+  String url;
+  bool isAllowed;
+
+  AddFundVendorModel(this.name, this.url, this.isAllowed);
+
+  AddFundVendorModel copyWith({String url, bool isAllowed}) {
+    return new AddFundVendorModel(this.name, url = this.url, isAllowed = isAllowed ?? this.isAllowed);
+  }
+}
